backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
nltk
sentence-transformers
faiss-cpu
